+++ doc/WhatsNew.rst
@@ -9,6 +9,10 @@ Ver 5.0.1 (unreleased)
 - Improved error checking on contains_pts() method for canvas items
 - Removed references to distutils module, deprecated in python 3.12
 - Added default antialiasing to cairo renderer
+- Fixed drawing of dimension labels for rectangles specified in WCS
+  coordinates
+- Fixed drawing of Crosshair canvas object in opengl backend
+- Fixed drawing of XRange and YRange in non-data coordinates
 
 Ver 5.0.0 (2024-02-24)
 ======================